Now obsolete but working Z-order based `OctreeOffset` type
It implements Z-order iteration, even for non-cube bounds. I originally wanted to use this as storage, but then realized that it's pretty pointless to store it in this format.

The implementation was pretty involved and cost me a lot of pain and agony, so I'd rather not throw it away in case I need it again in the future, so here we are.