Replace `derive_where` with `educe`
The main reason was, that doc-comments were no longer being highlighted properly by vscode due to `derive_where` being a standalone attribute macro rather than a regular `#derive[...]` with attributes.

Although, if I would've known that `educe` puts trait bounds based on the types of the *fields* I might've used it from the get-go. It might have a bit more compile-time overhead, but it's a lot nicer to use actually.